abi: Publish solidity ABI in cargo crate
