+++ abi/Cargo.toml
@@ -1,0 +1,32 @@
+[package]
+name = "abi"
+version = "0.1.0"
+edition = "2021"
+
+[features]
+relayer-types = [
+    "dep:ark-bn254",
+    "dep:ark-ec",
+    "dep:ark-ff",
+    "dep:jf-primitives",
+    "dep:num-bigint",
+    "dep:renegade-circuit-types",
+    "dep:renegade-circuits",
+    "dep:renegade-common",
+    "dep:renegade-constants",
+]
+
+[dependencies]
+alloy = { version = "0.14", features = ["essentials"] }
+
+# === Relayer Deps === #
+ark-bn254 = { version = "0.4.0", optional = true }
+ark-ec = { version = "0.4.0", optional = true }
+ark-ff = { version = "0.4.0", optional = true }
+num-bigint = { version = "0.4", optional = true }
+jf-primitives = { git = "https://github.com/renegade-fi/mpc-jellyfish.git", optional = true }
+
+renegade-circuits = { package = "circuits", git = "https://github.com/renegade-fi/renegade", branch = "zksecurity-audit-4-21-25", optional = true }
+renegade-circuit-types = { package = "circuit-types", git = "https://github.com/renegade-fi/renegade", branch = "zksecurity-audit-4-21-25", optional = true }
+renegade-common = { package = "common", git = "https://github.com/renegade-fi/renegade", branch = "zksecurity-audit-4-21-25", optional = true }
+renegade-constants = { package = "constants", git = "https://github.com/renegade-fi/renegade", branch = "zksecurity-audit-4-21-25", optional = true }
