+++ integration/src/contracts/mod.rs
@@ -1,5 +1,3 @@
 //! Solidity types for Renegade contracts
 
-pub mod darkpool;
 pub mod erc20;
-pub mod type_conversion;
